biophysics-utilities/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reaction-kinetics-playground"
version = "0.1.0"
description = "A modular computational framework for constructing and simulating chemical reaction networks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Murari", email = "murari@example.com"}
]
keywords = ["chemical kinetics", "reaction networks", "ODE simulation", "systems biology", "computational chemistry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "sympy>=1.9",
    "matplotlib>=3.4.0",
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
streamlit = [
    "streamlit>=1.20.0",
    "plotly>=5.0.0",
]
performance = [
    "cython>=0.29.0",
    "numba>=0.55.0",
]
export = [
    "python-libsbml>=5.19.0",
]
all = [
    "reaction-kinetics-playground[dev,streamlit,performance,export]"
]

[project.scripts]
kinetics = "kinetics_playground.cli.main:main"

[project.urls]
Homepage = "https://github.com/murari/reaction-kinetics-playground"
Documentation = "https://github.com/murari/reaction-kinetics-playground/docs"
Repository = "https://github.com/murari/reaction-kinetics-playground"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=kinetics_playground --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
